pyproject.toml
[project]
name = "slivit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "scikit-image>=0.26.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.0",
    "vit-pytorch>=1.19.1",
    "wandb>=0.25.1",
    "pyyaml>=6.0",
]
